Password reset for managers and fixes
